+++ NEWS.md
@@ -1,3 +1,28 @@
+# PEMcollectr 0.0.2
+
+* added and tested functionality for tracklogs
+
+* validation pairs modules now checks for data uploaded, staged, and complete
+
+* added photo module to view and upload photos to postgres
+
+* added AOI and labels to map
+
+* added ability to view staged and completed data on map
+
+* added ability to view uploaded data in map
+
+* added pending for transect triangles for sample areas that have one or 
+more in the staging table
+
+* unique constraint has been changed to transect_id and observer in case 
+sample plan is duplicated to measure an area twice
+
+* added a warning message for filtered data points that are already in the 
+database
+
+* incidentals will have a unique id generated based on point geometry
+
 # PEMcollectr 0.0.1
 
 * added ability to display all correctly uploaded track logs